requirements.txt
adal==1.2.1
aiounittest==1.3.0
attrs==19.3.0
Babel==2.7.0
botbuilder-core==4.8.0
botbuilder-dialogs==4.8.0
botbuilder-schema==4.8.0
botframework-connector==4.8.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
clickclick==1.2.2
connexion==2.6.0
cryptography==2.8
datedelta==1.3
emoji==0.5.4
Flask==1.1.1
grapheme==0.6.0
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpickle==1.2
jsonschema==3.2.0
MarkupSafe==1.1.1
msal==1.1.0
msrest==0.6.10
multipledispatch==0.6.0
oauthlib==3.1.0
openapi-spec-validator==0.2.8
pycparser==2.20
PyJWT==1.5.3
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
swagger-ui-bundle==0.0.6
urllib3==1.25.8
Werkzeug==1.0.1
